// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/BaseBalanceDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-agreements
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: Unknown version
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BaseBalanceDTO
 */
export interface BaseBalanceDTO {
    /**
     * 
     * @type {string}
     * @memberof BaseBalanceDTO
     */
    organisationId?: string;
    /**
     * 
     * @type {string}
     * @memberof BaseBalanceDTO
     */
    currency?: string;
    /**
     * 
     * @type {{ [key: string]: number; }}
     * @memberof BaseBalanceDTO
     */
    values?: { [key: string]: number; };
}

/**
 * Check if a given object implements the BaseBalanceDTO interface.
 */
export function instanceOfBaseBalanceDTO(value: object): value is BaseBalanceDTO {
    return true;
}

export function BaseBalanceDTOFromJSON(json: any): BaseBalanceDTO {
    return BaseBalanceDTOFromJSONTyped(json, false);
}

export function BaseBalanceDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): BaseBalanceDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'organisationId': json['organisationId'] == null ? undefined : json['organisationId'],
        'currency': json['currency'] == null ? undefined : json['currency'],
        'values': json['values'] == null ? undefined : json['values'],
    };
}

export function BaseBalanceDTOToJSON(json: any): BaseBalanceDTO {
    return BaseBalanceDTOToJSONTyped(json, false);
}

export function BaseBalanceDTOToJSONTyped(value?: BaseBalanceDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'organisationId': value['organisationId'],
        'currency': value['currency'],
        'values': value['values'],
    };
}